bigint: Minor cleanup, use `size_t`
- lua-bigint: Fix typo in `internal_mul_bigint` where `prod &
  DIGIT_SHIFT` should have been `prod & DIGIT_MASK`
- Rename several reverse iterators using `size_t` to be clearer

+++ bigint/bigint.h
@@ -48,10 +48,10 @@ struct BigInt {
     BigInt_Digit *data;
 
     // How many digits are validly indexable in `data`.
-    int len;
+    size_t len;
 
     // How many digits are allocated for in `data`.
-    int cap;
+    size_t cap;
 
     BigInt_Sign sign;
